refactor: centralize spring configuration and remove module auto-configs
- Delete all module-specific `AutoConfiguration` classes and corresponding `AutoConfiguration.imports` files across all modules
- Add `@EnableJpaRepositories` and `@EntityScan` to `StonemarkApplication` to globally scan the root `pt.estga` package
- Relocate `MessageSource` and `RestTemplate` bean definitions to the centralized `ApplicationConfig` in the `boot` module
