--- bookmark/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,1 +1,0 @@
-pt.estga.bookmark.config.BookmarkAutoConfiguration
