+++ boot/src/main/java/pt/estga/boot/StonemarkApplication.java
@@ -2,8 +2,10 @@ package pt.estga.boot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.persistence.autoconfigure.EntityScan;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.ComponentScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
@@ -12,6 +14,8 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @EnableCaching
 @SpringBootApplication
 @ComponentScan(basePackages = "pt.estga")
+@EnableJpaRepositories(basePackages = "pt.estga")
+@EntityScan(basePackages = "pt.estga")
 public class StonemarkApplication {
 
 	public static void main(String[] args) {
