--- chatbot/src/main/java/pt/estga/chatbot/config/ChatbotAutoConfiguration.java
@@ -1,20 +1,0 @@
-package pt.estga.chatbot.config;
-
-import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.context.MessageSource;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.support.ReloadableResourceBundleMessageSource;
-
-@AutoConfiguration
-@ComponentScan("pt.estga.chatbot")
-public class ChatbotAutoConfiguration {
-
-    @Bean
-    public MessageSource messageSource() {
-        ReloadableResourceBundleMessageSource messageSource = new ReloadableResourceBundleMessageSource();
-        messageSource.setBasename("classpath:messages");
-        messageSource.setDefaultEncoding("UTF-8");
-        return messageSource;
-    }
-}
