--- file/src/main/java/pt/estga/file/config/FileAutoConfiguration.java
@@ -1,17 +1,0 @@
-package pt.estga.file.config;
-
-import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.boot.persistence.autoconfigure.EntityScan;
-import org.springframework.cache.annotation.EnableCaching;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-import org.springframework.scheduling.annotation.EnableAsync;
-
-@AutoConfiguration
-@ComponentScan("pt.estga.file")
-@EnableJpaRepositories("pt.estga.file.repositories")
-@EntityScan("pt.estga.file.entities")
-@EnableAsync
-@EnableCaching
-public class FileAutoConfiguration {
-}
