--- notification/src/main/java/pt/estga/notification/NotificationAutoConfiguration.java
@@ -1,9 +1,0 @@
-package pt.estga.notification;
-
-import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.context.annotation.ComponentScan;
-
-@AutoConfiguration
-@ComponentScan("pt.estga.notification")
-public class NotificationAutoConfiguration {
-}
