--- shared-web/src/main/java/pt/estga/sharedweb/config/SharedWebAutoConfiguration.java
@@ -1,9 +1,0 @@
-package pt.estga.sharedweb.config;
-
-import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.context.annotation.ComponentScan;
-
-@AutoConfiguration
-@ComponentScan("pt.estga.sharedweb")
-public class SharedWebAutoConfiguration {
-}
