--- vision/src/main/java/pt/estga/vision/config/VisionAutoConfiguration.java
@@ -1,16 +1,0 @@
-package pt.estga.vision.config;
-
-import org.springframework.boot.autoconfigure.AutoConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.web.client.RestTemplate;
-
-@AutoConfiguration
-@ComponentScan("pt.estga.vision")
-public class VisionAutoConfiguration {
-
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
-}
